Split patron pin API integration tests into separate class MODUSERS-309
