+++ src/test/java/org/folio/rest/impl/PatronPinAPIIT.java
@@ -1,0 +1,124 @@
+
+package org.folio.rest.impl;
+
+import static org.hamcrest.CoreMatchers.is;
+
+import java.net.URI;
+
+import org.folio.postgres.testing.PostgresTesterContainer;
+import org.folio.rest.persist.PostgresClient;
+import org.folio.rest.tools.utils.NetworkUtils;
+import org.folio.support.User;
+import org.folio.support.VertxModule;
+import org.folio.support.http.FakeTokenGenerator;
+import org.folio.support.http.OkapiHeaders;
+import org.folio.support.http.PatronPinClient;
+import org.folio.support.http.UsersClient;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import io.vertx.core.Vertx;
+import io.vertx.junit5.VertxExtension;
+import io.vertx.junit5.VertxTestContext;
+import lombok.SneakyThrows;
+
+@ExtendWith(VertxExtension.class)
+class PatronPinAPIIT {
+  private static UsersClient usersClient;
+  private static PatronPinClient patronPinClient;
+
+  @BeforeAll
+  @SneakyThrows
+  static void beforeAll(Vertx vertx, VertxTestContext context) {
+    final var tenant = "patron_pin_integration_tests";
+    final var token = new FakeTokenGenerator().generateToken();
+
+    PostgresClient.setPostgresTester(new PostgresTesterContainer());
+
+    final var port = NetworkUtils.nextFreePort();
+
+    final var headers = new OkapiHeaders("http://localhost:" + port,
+      tenant, token);
+
+    usersClient = new UsersClient(new URI("http://localhost:" + port), headers);
+    patronPinClient = new PatronPinClient(new URI("http://localhost:" + port), headers);
+
+    final var module = new VertxModule(vertx);
+
+    module.deployModule(port)
+      .onComplete(context.succeeding(res -> module.enableModule(headers,
+          true, false)
+        .onComplete(context.succeedingThenComplete())));
+  }
+
+  @BeforeEach
+  public void beforeEach() {
+    usersClient.deleteAllUsers();
+  }
+
+  @ParameterizedTest
+  @ValueSource(strings = {"1468", "ThisIsALonger1234PinWithSomeNumbers", "7778"})
+  void canVerifyCorrectPatronPin(String pin) {
+    final var user = usersClient.createUser("apple");
+
+    patronPinClient.assignPatronPin(user.getId(), pin);
+
+    enteredPinIsValid(user, pin);
+  }
+
+  @ParameterizedTest
+  @CsvSource({"1468,1467", "ThisIsALonger1234PinWithSomeNumbers,1111"})
+  void canVerifyIncorrectPatronPin(String actualPin, String attemptedPin) {
+    final var user = usersClient.createUser("apple");
+
+    patronPinClient.assignPatronPin(user.getId(), actualPin);
+
+    enteredPinIsInvalid(user, attemptedPin);
+  }
+
+  @Test
+  void canVerifyPatronPinForUserWithNoPinAssigned() {
+    final var user = usersClient.createUser("apple");
+
+    enteredPinIsInvalid(user, "1234");
+  }
+
+  @Test
+  void canReassignPatronPin() {
+    final var user = usersClient.createUser("apple");
+
+    patronPinClient.assignPatronPin(user.getId(), "1234");
+
+    patronPinClient.assignPatronPin(user.getId(), "4567");
+
+    enteredPinIsValid(user, "4567");
+
+    enteredPinIsInvalid(user, "1234");
+  }
+
+  @Test
+  void canRemovePatronPin() {
+    final var user = usersClient.createUser("apple");
+
+    patronPinClient.assignPatronPin(user.getId(), "1234");
+
+    patronPinClient.removePatronPin(user.getId());
+
+    enteredPinIsInvalid(user, "1234");
+  }
+
+  private void enteredPinIsValid(User user, String pin) {
+    patronPinClient.verifyPatronPin(user.getId(), pin)
+      .statusCode(is(200));
+  }
+
+  private void enteredPinIsInvalid(User user, String pin) {
+    patronPinClient.verifyPatronPin(user.getId(), pin)
+      .statusCode(is(422));
+  }
+}
