+++ src/test/java/org/folio/support/http/OkapiHeaders.java
@@ -3,8 +3,7 @@ package org.folio.support.http;
 import lombok.Value;
 
 @Value
-public
-class OkapiHeaders {
+public class OkapiHeaders {
   String okapiUrl;
   String tenantId;
   String token;
